requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
joblib>=1.1.0
Pillow>=9.0.0
flask
tensorflow
protobuf